core/worker/requirements.txt
fastapi
uvicorn[standard]
aio-pika
python-dotenv
httpx
pydantic-settings
SQLAlchemy
asyncpg
python-docx
langchain